session: move tab naming helpers out of instance
