+++ scripts/file-length-allowlist.txt
@@ -19,7 +19,7 @@
 # --- production code (limit 1000) ---
 daemon/control.go 2587
 app/app.go 1867
-session/instance.go 1531
+session/instance.go 1400
 session/tmux/tmux.go 1393
 ui/sidebar.go 1258
 config/config.go 1128
